+++ src/Stmt.c
@@ -13,18 +13,27 @@ static ReturnResult visitExpressionStmt(StmtVisitor* visitor,Stmt* stmt){
     ReturnResult r; r.value.string=NULL; return r;
 }
 void new_Expression (Expression * inObj,Expr* expressionparam){
-	inObj->super.vtable.accept = &acceptExpression;
-	inObj->super.vtable.delete_Stmt = &delete_Expression ;
+    inObj->super.vtable.accept = &acceptExpression;
+    
+    inObj->super.vtable.delete_Stmt = &delete_Expression ;
 	inObj->expression = expressionparam;
+	inObj->super.owner_references=1;
+	inObj->super.id = addtoStmtCounter();
 }
 void delete_Expression (Stmt* arg){
 	Expression * expr = (Expression *)arg;
 
+if(expr->super.owner_references ==1){
     delete_Expr(expr->expression);
     expr->expression=NULL;
 
 	free(expr);
 	expr=NULL;
+}
+
+    else{
+        expr->super.owner_references--;
+    }
 }
 
 ReturnResult acceptExpression(Stmt *arg, StmtVisitor* visitor){
@@ -37,18 +46,27 @@ static ReturnResult visitPrintStmt(StmtVisitor* visitor,Stmt* stmt){
     ReturnResult r; r.value.string=NULL; return r;
 }
 void new_Print (Print * inObj,Expr* expressionparam){
-	inObj->super.vtable.accept = &acceptPrint;
-	inObj->super.vtable.delete_Stmt = &delete_Print ;
+    inObj->super.vtable.accept = &acceptPrint;
+    
+    inObj->super.vtable.delete_Stmt = &delete_Print ;
 	inObj->expression = expressionparam;
+	inObj->super.owner_references=1;
+	inObj->super.id = addtoStmtCounter();
 }
 void delete_Print (Stmt* arg){
 	Print * expr = (Print *)arg;
 
+if(expr->super.owner_references ==1){
     delete_Expr(expr->expression);
     expr->expression=NULL;
 
 	free(expr);
 	expr=NULL;
+}
+
+    else{
+        expr->super.owner_references--;
+    }
 }
 
 ReturnResult acceptPrint(Stmt *arg, StmtVisitor* visitor){
@@ -61,14 +79,18 @@ static ReturnResult visitVarStmt(StmtVisitor* visitor,Stmt* stmt){
     ReturnResult r; r.value.string=NULL; return r;
 }
 void new_Var (Var * inObj,Token* nameparam,Expr* initializerparam){
-	inObj->super.vtable.accept = &acceptVar;
-	inObj->super.vtable.delete_Stmt = &delete_Var ;
+    inObj->super.vtable.accept = &acceptVar;
+    
+    inObj->super.vtable.delete_Stmt = &delete_Var ;
 	inObj->name = nameparam;
 	inObj->initializer = initializerparam;
+	inObj->super.owner_references=1;
+	inObj->super.id = addtoStmtCounter();
 }
 void delete_Var (Stmt* arg){
 	Var * expr = (Var *)arg;
 
+if(expr->super.owner_references ==1){
     delete_Token(expr->name);
     expr->name=NULL;
     delete_Expr(expr->initializer);
@@ -78,9 +100,28 @@ void delete_Var (Stmt* arg){
 	expr=NULL;
 }
 
+    else{
+        expr->super.owner_references--;
+    }
+}
+
 ReturnResult acceptVar(Stmt *arg, StmtVisitor* visitor){
     return visitor->vtable.visitVarStmt(visitor,arg);
 }
 void delete_Stmt(Stmt* expr){
 	expr->vtable.delete_Stmt(expr);
 }
+short int addtoStmtCounter(void){
+     Stmtcounter += 1;
+     return Stmtcounter;
+    }
+
+Stmt* getStmtReference(Stmt* arg){
+    arg->owner_references += 1;
+    return arg;
+}
+void removeStmtReference(Stmt* arg){
+    if(arg->owner_references > 1)
+        arg->owner_references -= 1;
+}
+
